[leetcode/ac] 662. Maximum Width of Binary Tree
Runtime: 8 ms, faster than 78.06% of C++ online submissions for Maximum Width of Binary Tree.
Memory Usage: 17.2 MB, less than 14.30% of C++ online submissions for Maximum Width of Binary Tree.

+++ leetcode/solved/662_Maximum_Width_of_Binary_Tree.cc
@@ -12,11 +12,11 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-        map<int, pair<int,int>> widths;
+        map<int, pair<unsigned long long, unsigned long long>> widths;
         
         widthOfBinaryTree(root, 0, 1, widths);
         
-        int ans = 0;
+        unsigned long long ans = 0;
         for(const auto &e : widths) {
             ans = max(ans, e.second.second - e.second.first + 1);
         }
@@ -26,8 +26,8 @@ public:
     
 private:
     void widthOfBinaryTree(const TreeNode *root,
-                           const int depth, const int idx,
-                           map<int, pair<int,int>> &widths)
+                           const int depth, const unsigned long long idx,
+                           map<int, pair<unsigned long long, unsigned long long>> &widths)
     {
         if(root == nullptr) return;
